+++ examples/resources/pingdirectory_third_party_search_entry_criteria/import.sh
@@ -1,0 +1,2 @@
+# "thirdPartySearchEntryCriteriaId" should be the id of the Third Party Search Entry Criteria to be imported
+terraform import pingdirectory_third_party_search_entry_criteria.myThirdPartySearchEntryCriteria thirdPartySearchEntryCriteriaId